fix(cli): route channel/connection commands through consolidated print.ts
main added the `channel` module (and `connection update`) importing
printServiceError from agent/commands/errors.js. This branch's refactor
dropped that re-export shim and moved printServiceError to shared/trpc/print.ts,
so after the rebase those imports no longer resolved. Repoint them at print.ts
— keeping the resolve helpers (exitCodeForResolveError, printResolveError) from
errors.js — so every command group keeps a single source of error printing.
