+++ packages/cli/src/modules/connection/commands/update.ts
@@ -1,6 +1,6 @@
 import { cancel, isCancel } from "@clack/prompts";
 import { Command } from "commander";
-import { printServiceError } from "../../agent/commands/errors.js";
+import { printServiceError } from "../../shared/trpc/print.js";
 import type { CompatService, ConfigService } from "../../cli/index.js";
 import {
   EXIT_BELOW_FLOOR,
